pk记录查询显示完成

+++ capacity_parent/capacity_front/src/main/java/com/zhidejiaoyu/student/business/shipconfig/service/impl/ShipTestServiceImpl.java
@@ -99,7 +99,7 @@ public class ShipTestServiceImpl extends BaseServiceImpl<StudentMapper, Student>
         shipAddEquipmentService.updateLeaderBoards(student);
         Long currentRanking = this.getCurrentRanking(student);
         if (currentRanking < expansion.getRanking() && currentRanking != 0) {
-            expansion.setRanking(currentRanking.intValue());
+            expansion.setRanking(currentRanking);
         }
         studentExpansionMapper.updateById(expansion);
         returnMap.put("status", 1);
